requirements.txt
# Core Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.10.1

# Machine Learning
xgboost==1.7.6
lightgbm==3.3.5
catboost==1.2.5          
optuna==3.5.0

# Imbalanced Learning
imbalanced-learn==0.11.0

# NLP & Text Processing
nltk==3.8.1
textblob==0.17.1
langdetect==1.0.9
spacy==3.7.2
emoji==2.8.0
textstat==0.7.3
contractions==0.1.73
wordcloud==1.9.2

# Web Scraping & Data Collection
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
praw==7.7.0
python-whois==0.8.0
tldextract==5.1.1

# API Development
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database (optional)
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities
joblib==1.3.2
loguru==0.7.2
tqdm==4.66.1
pyyaml==6.0.1

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Jupyter (for exploration)
ipykernel==6.25.0
jupyter==1.0.0

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.18.0

# Email Validation
email-validator==2.1.0
validate-email==1.3

# URL Analysis
urlextract==1.8.0
tld==0.13

# Grammar checking
language-tool-python==2.7.1
feedparser
